refactor: seperate the upload function
